guest functionality added, and clear cart when you log out

+++ client/store/reducers/shoppingCart.js
@@ -3,9 +3,10 @@ import axios from "axios";
 // ACTION TYPES
 
 const GET_SHOPPING_CART = "GET_SHOPPING_CART";
+const UPDATE_ITEM_IN_CART = "UPDATE_ITEM_IN_CART";
 const DELETE_ITEM_IN_CART = "DELETE_ITEM_IN_CART";
-const ADD_ITEM_TO_CART_AS_GUEST = "ADD_ITEM_TO_CART_AS_GUEST";
-const UPDATE_ITEM_AS_GUEST = 'UPDATE_ITEM_AS_GUEST'
+const ADD_ITEM_TO_CART = "ADD_ITEM_TO_CART";
+const CLEAR_CART = 'CLEAR_CART'
 
 // ACTION CREATORS
 
@@ -23,19 +24,26 @@ const deleteItemFromCart = itemId => {
   };
 };
 
-const addItemToCartAsGuest = (item, quantity) => {
+const updateItemFromCart = (itemId, quantity) => {
   return {
-    type: ADD_ITEM_TO_CART_AS_GUEST,
+    type: UPDATE_ITEM_IN_CART,
+    itemId,
+    quantity,
+  };
+};
+
+const addItemToCart = (item, quantity) => {
+  return {
+    type: ADD_ITEM_TO_CART,
     item,
     quantity,
   };
 };
 
-const updateItemAsGuest = (itemId, quantity) => {
+export const clearCart = () => {
   return {
-    type: UPDATE_ITEM_AS_GUEST,
-    itemId,
-    quantity,
+    type: CLEAR_CART,
+    cart: []
   }
 }
 
@@ -62,8 +70,10 @@ export const postItemToCart = (id, amount, history) => async dispatch => {
   try {
     const token = window.localStorage.getItem("token");
 
+    const { data: item } = await axios.get(`/api/products/${id}`);
+
     if (token) {
-      await axios.post(
+      const { data: cart } = await axios.post(
         `/api/shoppingCart/${id}`,
         { quantity: amount },
         {
@@ -73,11 +83,13 @@ export const postItemToCart = (id, amount, history) => async dispatch => {
         }
       );
 
-      fetchShoppingCart();
-    } else {
-      const { data: item } = await axios.get(`/api/products/${id}`);
+      item.shoppingCart = cart.quantity;
 
-      dispatch(addItemToCartAsGuest(item, amount));
+      dispatch(addItemToCart(item, amount));
+    } else {
+      item.shoppingCart = { quantity: amount };
+
+      dispatch(addItemToCart(item, amount));
     }
 
     history.push("/shoppingCart");
@@ -100,11 +112,8 @@ export const updateItemInCart = (id, amount) => async dispatch => {
           },
         }
       );
-
-      fetchShoppingCart();
-    } else {
-      dispatch(updateItemAsGuest(id, amount))
     }
+    dispatch(updateItemFromCart(id, amount));
   } catch (error) {
     console.log(error);
   }
@@ -114,11 +123,13 @@ export const deleteItemInCart = id => async dispatch => {
   try {
     const token = window.localStorage.getItem("token");
 
-    await axios.delete(`/api/shoppingCart/${id}`, {
-      headers: {
-        authorization: token,
-      },
-    });
+    if (token) {
+      await axios.delete(`/api/shoppingCart/${id}`, {
+        headers: {
+          authorization: token,
+        },
+      });
+    }
 
     dispatch(deleteItemFromCart(id));
   } catch (error) {
@@ -138,26 +149,28 @@ const reducer = (state = initialState, action) => {
       return action.cart;
     case DELETE_ITEM_IN_CART:
       return [...state.filter(product => product.id !== action.itemId)];
-    case ADD_ITEM_TO_CART_AS_GUEST:
+    case ADD_ITEM_TO_CART:
       const newItem = {
         ...action.item,
         shoppingCart: { quantity: action.quantity },
       };
+
       return [...state, newItem];
-    case UPDATE_ITEM_AS_GUEST:
+    case UPDATE_ITEM_IN_CART:
       const newState = state.map(product => {
         if (product.id === action.itemId) {
-          const newProduct = {...product}
+          const newProduct = { ...product };
 
-          newProduct.shoppingCart = {quantity: Number(action.quantity)}
+          newProduct.shoppingCart = { quantity: Number(action.quantity) };
 
-          return newProduct
+          return newProduct;
         } else {
-          return product
+          return product;
         }
-      })
-
-      return newState
+      });
+      return newState;
+    case CLEAR_CART:
+      return action.cart
     default:
       return state;
   }
